// packages/defi_orders_frontend/codegen/UniversalSwap.ts
export type AssetStructOutput = [
  string,
  string,
  BigNumber,
  BigNumber,
  string
] & {
  pool: string;
  manager: string;
  tokenId: BigNumber;
  liquidity: BigNumber;
  data: string;
};

export type ProvidedStruct = {
  tokens: PromiseOrValue<string>[];
  amounts: PromiseOrValue<BigNumberish>[];
  nfts: AssetStruct[];
};

export type ProvidedStructOutput = [
  string[],
  BigNumber[],
  AssetStructOutput[]
] & { tokens: string[]; amounts: BigNumber[]; nfts: AssetStructOutput[] };